Add support for Yahoo! JAPAN AMP cache
This is similar to Google’s AMP viewer. This doesn’t work with the generic AMP handler because the AMP page is shown in an iframe, which we can’t access.

It might be possible to inject the generic handler in all frames but that could cause issues on other websites, if they embedded an AMP page with an iframe.